@@ -2,69 +2,70 @@
   "name": "teallarson-dev",
   "version": "1.0.0",
   "private": true,
+  "engines": {
+    "node": ">=22.0.0"
+  },
   "scripts": {
     "start": "cross-env SOCKET=true node ./scripts/next-remote-watch.js ./data",
-    "dev": "next dev",
+    "dev": "next dev --turbopack",
     "build": "next build && node ./scripts/generate-sitemap",
     "serve": "next start",
     "analyze": "cross-env ANALYZE=true next build",
-    "lint": "next lint --fix --dir pages --dir components --dir lib --dir layouts --dir scripts"
+    "lint": "next lint --fix --dir app --dir pages --dir components --dir lib --dir layouts --dir scripts"
   },
   "dependencies": {
-    "@fontsource/inter": "4.5.2",
-    "@tailwindcss/forms": "^0.4.0",
+    "@fontsource/inter": "^5.0.0",
+    "@tailwindcss/forms": "^0.5.0",
     "@tailwindcss/typography": "^0.5.0",
     "autoprefixer": "^10.4.0",
-    "esbuild": "^0.13.13",
-    "github-slugger": "^1.3.0",
-    "gray-matter": "^4.0.2",
-    "image-size": "1.0.0",
-    "mdx-bundler": "^8.0.0",
-    "next": "12.0.9",
-    "next-themes": "^0.0.14",
-    "next-use-posthog": "^1.16.0",
-    "postcss": "^8.4.5",
-    "preact": "^10.6.2",
-    "react": "17.0.2",
-    "react-dom": "17.0.2",
-    "reading-time": "1.3.0",
-    "rehype-autolink-headings": "^6.1.0",
+    "esbuild": "^0.20.0",
+    "github-slugger": "^2.0.0",
+    "gray-matter": "^4.0.3",
+    "image-size": "^1.0.2",
+    "js-yaml": "^4.1.1",
+    "katex": "^0.16.25",
+    "mdast-util-to-string": "^3.2.0",
+    "mdx-bundler": "^10.0.0",
+    "next": "^16.0.3",
+    "next-themes": "^0.2.1",
+    "postcss": "^8.4.0",
+    "posthog-js": "^1.0.0",
+    "react": "^18.3.1",
+    "react-dom": "^18.3.1",
+    "reading-time": "^1.5.0",
+    "rehype-autolink-headings": "^7.1.0",
     "rehype-citation": "^0.2.0",
-    "rehype-katex": "^6.0.2",
-    "rehype-preset-minify": "6.0.0",
+    "rehype-katex": "^7.0.0",
     "rehype-prism-plus": "^1.1.3",
-    "rehype-slug": "^5.0.0",
-    "remark-footnotes": "^4.0.1",
-    "remark-gfm": "^3.0.1",
-    "remark-math": "^5.1.1",
-    "sharp": "^0.28.3",
-    "smoothscroll-polyfill": "^0.4.4",
-    "tailwindcss": "^3.0.18",
-    "unist-util-visit": "^4.0.0"
+    "rehype-slug": "^6.0.0",
+    "remark-gfm": "^4.0.0",
+    "remark-math": "^6.0.0",
+    "sharp": "^0.33.0",
+    "tailwindcss": "^3.4.0",
+    "unist": "npm:@types/unist@^3.0.3",
+    "unist-util-visit": "^5.0.0",
+    "vfile": "^6.0.3"
   },
   "devDependencies": {
-    "@next/bundle-analyzer": "12.0.9",
-    "@svgr/webpack": "^6.1.2",
-    "@types/react": "^17.0.14",
-    "@types/tailwindcss": "^2.2.0",
-    "@typescript-eslint/eslint-plugin": "^4.28.1",
-    "@typescript-eslint/parser": "^4.28.1",
+    "@next/bundle-analyzer": "^16.0.3",
+    "@svgr/webpack": "^8.1.0",
+    "@types/node": "^20.0.0",
+    "@types/react": "^18.3.0",
+    "@types/react-dom": "^18.3.0",
+    "@types/unist": "^3.0.3",
     "cross-env": "^7.0.3",
     "dedent": "^0.7.0",
-    "eslint": "^7.29.0",
-    "eslint-config-next": "12.0.9",
-    "eslint-config-prettier": "^8.3.0",
-    "eslint-plugin-prettier": "^3.3.1",
-    "file-loader": "^6.0.0",
-    "globby": "11.0.3",
-    "inquirer": "^8.1.1",
-    "lint-staged": "^11.0.0",
+    "eslint-config-next": "^16.0.3",
+    "eslint-config-prettier": "^9.1.0",
+    "globby": "^14.0.0",
+    "inquirer": "^9.2.0",
+    "lint-staged": "^15.0.0",
     "next-remote-watch": "^1.0.0",
-    "prettier": "^2.5.1",
-    "prettier-plugin-tailwindcss": "^0.1.4",
-    "socket.io": "^4.4.0",
-    "socket.io-client": "^4.4.0",
-    "typescript": "^4.6.1-rc"
+    "prettier": "^3.2.0",
+    "prettier-plugin-tailwindcss": "^0.5.0",
+    "socket.io": "^4.7.0",
+    "socket.io-client": "^4.7.0",
+    "typescript": "^5.3.0"
   },
   "lint-staged": {
     "*.+(js|jsx|ts|tsx)": [
