@@ -1,4 +1,3 @@
-import { TagSEO } from '@/components/SEO'
 import siteMetadata from '@/data/siteMetadata'
 import ListLayout from '@/layouts/ListLayout'
 import generateRss from '@/lib/generate-rss'
@@ -6,35 +5,53 @@ import { getAllFilesFrontMatter } from '@/lib/mdx'
 import { getAllTags } from '@/lib/tags'
 import kebabCase from '@/lib/utils/kebabCase'
 import fs from 'fs'
-import { GetStaticProps, InferGetStaticPropsType } from 'next'
 import path from 'path'
 import { PostFrontMatter } from 'types/PostFrontMatter'
+import { Metadata } from 'next'
+import { notFound } from 'next/navigation'
 
 const root = process.cwd()
 
-export async function getStaticPaths() {
+export async function generateStaticParams() {
   const tags = await getAllTags('blog')
+  return Object.keys(tags).map((tag) => ({
+    tag,
+  }))
+}
+
+export async function generateMetadata({
+  params,
+}: {
+  params: Promise<{ tag: string }>
+}): Promise<Metadata> {
+  const { tag } = await params
+  const tags = await getAllTags('blog')
+  if (!tags[tag]) {
+    notFound()
+  }
 
   return {
-    paths: Object.keys(tags).map((tag) => ({
-      params: {
-        tag,
-      },
-    })),
-    fallback: false,
+    title: `${tag} - ${siteMetadata.title}`,
+    description: `${tag} tags - ${siteMetadata.author}`,
   }
 }
 
-export const getStaticProps: GetStaticProps<{ posts: PostFrontMatter[]; tag: string }> = async (
-  context
-) => {
-  const tag = context.params.tag as string
+export default async function TagPage({
+  params,
+}: {
+  params: Promise<{ tag: string }>
+}) {
+  const { tag } = await params
   const allPosts = await getAllFilesFrontMatter('blog')
   const filteredPosts = allPosts.filter(
     (post) => post.draft !== true && post.tags.map((t) => kebabCase(t)).includes(tag)
   )
 
-  // rss
+  if (filteredPosts.length === 0) {
+    notFound()
+  }
+
+  // Generate RSS
   if (filteredPosts.length > 0) {
     const rss = generateRss(filteredPosts, `tags/${tag}/feed.xml`)
     const rssPath = path.join(root, 'public', 'tags', tag)
@@ -42,19 +59,9 @@ export const getStaticProps: GetStaticProps<{ posts: PostFrontMatter[]; tag: str
     fs.writeFileSync(path.join(rssPath, 'feed.xml'), rss)
   }
 
-  return { props: { posts: filteredPosts, tag } }
-}
-
-export default function Tag({ posts, tag }: InferGetStaticPropsType<typeof getStaticProps>) {
   // Capitalize first letter and convert space to dash
   const title = tag[0].toUpperCase() + tag.split(' ').join('-').slice(1)
-  return (
-    <>
-      <TagSEO
-        title={`${tag} - ${siteMetadata.title}`}
-        description={`${tag} tags - ${siteMetadata.author}`}
-      />
-      <ListLayout posts={posts} title={title} />
-    </>
-  )
+
+  return <ListLayout posts={filteredPosts} title={title} />
 }
+
