@@ -1,5 +1,7 @@
 import NextImage, { ImageProps } from 'next/image'
 
-const Image = ({ ...rest }: ImageProps) => <NextImage {...rest} />
+const Image = ({ loading = 'lazy', ...rest }: ImageProps) => (
+  <NextImage loading={loading} {...rest} />
+)
 
 export default Image
