@@ -14,13 +14,19 @@ interface Props {
 const LayoutWrapper = ({ children }: Props) => {
   return (
     <SectionContainer>
-      <div className="flex h-screen flex-col justify-between">
+      <a
+        href="#main-content"
+        className="sr-only focus:not-sr-only focus:absolute focus:left-4 focus:top-4 focus:z-50 focus:rounded-md focus:bg-primary-600 focus:px-4 focus:py-2 focus:text-white focus:outline-none focus:ring-2 focus:ring-primary-500 focus:ring-offset-2"
+      >
+        Skip to main content
+      </a>
+      <div className="flex min-h-screen flex-col justify-between">
         <header className="flex items-center justify-between py-10">
           <div>
             <Link href="/" aria-label={siteMetadata.headerTitle}>
               <div className="flex items-center justify-between">
                 {typeof siteMetadata.headerTitle === 'string' ? (
-                  <div className="hidden h-6 text-2xl font-semibold sm:block">
+                  <div className="gradient-text-viewport hidden h-6 text-2xl font-semibold sm:block">
                     {siteMetadata.headerTitle}
                   </div>
                 ) : (
@@ -35,7 +41,7 @@ const LayoutWrapper = ({ children }: Props) => {
                 <Link
                   key={link.title}
                   href={link.href}
-                  className="p-1 font-medium text-gray-900 dark:text-gray-100 sm:p-4"
+                  className="bg-gradient-to-r from-gray-900 via-gray-700 to-gray-900 bg-clip-text p-1 font-medium text-transparent transition-all hover:from-primary-500 hover:via-primary-500 hover:to-primary-500 hover:bg-clip-text dark:from-gray-100 dark:via-gray-300 dark:to-gray-100 dark:hover:from-primary-400 dark:hover:via-primary-400 dark:hover:to-primary-400 sm:p-4"
                 >
                   {link.title}
                 </Link>
@@ -45,7 +51,9 @@ const LayoutWrapper = ({ children }: Props) => {
             <MobileNav />
           </div>
         </header>
-        <main className="mb-auto">{children}</main>
+        <main id="main-content" className="mb-auto" tabIndex={-1}>
+          {children}
+        </main>
         <Footer />
       </div>
     </SectionContainer>
