@@ -1,3 +1,5 @@
+'use client'
+
 import { useEffect, useState } from 'react'
 
 const ScrollTopAndComment = () => {
@@ -17,7 +19,10 @@ const ScrollTopAndComment = () => {
     window.scrollTo({ top: 0, behavior: 'smooth' })
   }
   const handleScrollToComment = () => {
-    document.getElementById('comment').scrollIntoView()
+    const commentElement = document.getElementById('comment')
+    if (commentElement) {
+      commentElement.scrollIntoView()
+    }
   }
   return (
     <div
