@@ -1,5 +1,25 @@
 const talksData = [
   {
+    title: 'Staying Compliant in the World of AI | Move(data) 2025',
+    date: '03-01-2025', 
+    description: 'Join Teal Larson, a software engineer at Airbyte, as she delves into staying compliant in AI using Airbyte mappings. Learn about the evolution of data movement from ETL to ELT, and how AI is transforming data processing for compliance and security. Discover how Airbyte mappings can help you hash, encrypt, and filter sensitive data before it reaches its destination, ensuring compliance without adding extra tools. Perfect for anyone looking to refine data movements in AI-driven projects.', // TODO: Update with actual description
+    imgSrc: '/static/images/talks/2025-03-20-staying-compliant-in-ai.png',
+    href: 'https://www.youtube.com/watch?v=M2ngk4K8-dU&t=238s',
+  },  
+  {
+    title: 'Open Source Powers the Modern Data Stack (WITCON Pgh 2022)', 
+    date: '11-09-2022', 
+    description: 'Open source powers the modern data stack. In this talk, I\'ll explore how open source software is driving innovation in the data space and how it\'s transforming the way we build and use data pipelines. (given live, no recording available 😞)', 
+  },
+  {
+    title: 'A Brief Talk About Nothing (Code & Supply Aug 2022 Lightning Talks)', 
+    date: '08-20-2022',
+    description: 'Exploring a few of the ways programming refers to "nothing", a quick hop/skip/jump through set theory, and why you can\'t coalesce with an empty set in SQL.', // TODO: Update with actual description
+    imgSrc: 'static/images/talks/2022-08-20-2022-talk-about-nothing.png', 
+    href: 'https://www.youtube.com/live/TxPcR3e4Wmk?si=ku60w42crw_wzhpm',
+ 
+  },
+  {
     title: 'From Teacher to Software Engineer: A Real-Life Success Story | Meet The Bytes',
     date: '08-11-2022',
     description:
@@ -15,13 +35,14 @@ const talksData = [
     href: 'https://www.youtube.com/watch?v=OsJVgfjanPI',
   },
   {
-    title: 'Code and Supply Lightning Talks December 2021: Why Use @use in your SASS',
+    title: 'Why Use @use in your SASS (Code & Supply Dec 2021 Lightning Talks)',
     date: '12-20-2021',
     description:
       'The development, compile-time, and runtime benefits of using @use and @forward over @import with modular SASS.',
     imgSrc: '/static/images/talks/2021-12-20-use-modular-sass.png',
     href: 'https://youtu.be/1UtklNrB8XA?t=2486',
   },
+ 
 ]
 
 export default talksData
